requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django
django-celery-beat==2.4.0
django-timezone-field==5.0
flower==1.2.0
humanize==4.4.0
kombu==5.2.4
prometheus-client==0.15.0
prompt-toolkit==3.0.36
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2022.7
redis==4.4.0
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tornado==6.2
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.5
django-enumchoicefield
